+++ src/pages/answer-page/AnswerPage.jsx
@@ -7,25 +7,15 @@ import FeedCardContainer from '@components/ui/molecules/feed-card/FeedCardContai
 import { getAnswerLists } from '@api/answer/getAnswerLists';
 import getUserData from '@api/getUserData';
 import { useAsyncOnMount } from '@hooks/useAsyncOnMount';
-import { useConfirmAlert } from '@hooks/useConfirmAlert';
 import useSetUser from '@hooks/useSetUser';
 import { useSNSShare } from '@hooks/useSNSShare';
 
 const AnswerPage = () => {
   const { copyUrl, shareToFacebook, shareToKakaotalk } = useSNSShare();
   const { userName, userProfile, createdAt, questionCount } = useSetUser(getUserData);
-  const { showConfirm, ConfirmAlertComponent } = useConfirmAlert();
   const [, , answerResults] = useAsyncOnMount(getAnswerLists);
 
   console.log(answerResults);
-
-  const handleConfirm = () => {
-    console.log('Confirmed!');
-  };
-
-  const handleCancel = () => {
-    console.log('Cancelled!');
-  };
 
   // feedcard type default가 null이다.
   // id 값이 주소에 있는 페이지라면? edit/reply
@@ -41,11 +31,6 @@ const AnswerPage = () => {
         </StSnsWrapper>
       </StQuestFeedPageWrapper>
       <FeedCardContainer cardLength={questionCount} answerResults={answerResults?.results} />
-      <Button onClickHandler={() => showConfirm(handleConfirm, handleCancel)}>전체 피드 삭제</Button>
-      <ConfirmAlertComponent
-        title='전체 피드가 삭제됩니다'
-        content={`피드를 모두 삭제하시겠어요? \n 삭제된 피드는 복구할 수 없습니다.`}
-      />
     </div>
   );
 };
